+++ packages/terra-form-select/src/SingleSelectField.jsx
@@ -35,6 +35,11 @@ const propTypes = {
    */
   help: PropTypes.node,
   /**
+   * @private
+   * Whether to hide the required indicator on the label.
+   */
+  hideRequired: PropTypes.bool,
+  /**
    * Whether the field displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
    */
   isIncomplete: PropTypes.bool,
@@ -46,6 +51,11 @@ const propTypes = {
    * Whether the field displays as Invalid. Use when value does not meet validation pattern.
    */
   isInvalid: PropTypes.bool,
+  /**
+   * @private
+   * Whether the label is hidden. Allows hiding the label while meeting accessibility guidelines.
+   */
+  isLabelHidden: PropTypes.bool,
   /**
    * Additional attributes to spread onto the label.
    */
@@ -61,6 +71,7 @@ const propTypes = {
    */
   maxWidth: PropTypes.string,
   /**
+   * @private
    * Callback function triggered when the select value changes. function(value)
    */
   onChange: PropTypes.func,