+++ packages/terra-form-select/src/shared/_Option.jsx
@@ -94,6 +94,12 @@ const Option = ({
     }
   }
 
+  let label = display;
+  if (customProps.index && customProps.totalOptions) {
+    label = `${display} (${customProps.index} of ${customProps.totalOptions})`;
+  }
+  const itemLabel = isSelected || customProps.index === 1 ? `${customProps.expandedStateText} ${label}` : label;
+
   return (
     <li
       role={role}
@@ -105,7 +111,7 @@ const Option = ({
       aria-disabled={disabled}
       tabIndex="0" // eslint-disable-line jsx-a11y/no-noninteractive-tabindex
       data-terra-select-option
-      aria-label={`${display} (${customProps.index} of ${customProps.totalOptions})`}
+      aria-label={itemLabel}
     >
       {(isCheckable || isAddOption) && <span className={cx('icon')} />}
       <span className={cx('display')}>{display}</span>