+++ packages/terra-form-select/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added `aria-label` to list options.
+
+* Changed
+  * Changed `hideRequired`, `isLabelHidden`, `onChange` props to private.
+  * Updated aria attributes.
+
+* Fixed
+  * Screen reader response for single-select.
 
 ## 6.39.0 - (February 15, 2023)